requirements.txt
appdirs==1.4.4
attrs==22.1.0
autopage==0.5.1
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cliff==4.1.0
cmd2==2.4.2
cryptography==38.0.4
debtcollector==2.5.0
decorator==5.1.1
dogpile.cache==1.1.8
idna==3.4
importlib-metadata==5.1.0
iso8601==1.1.0
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
keystoneauth1==5.1.0
mergedeep==1.3.4
msgpack==1.0.4
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==0.103.0
os-service-types==1.7.0
osc-lib==2.6.2
oslo.config==9.0.0
oslo.i18n==5.1.0
oslo.serialization==5.0.0
oslo.utils==6.1.0
packaging==21.3
paramiko==2.12.0
pbr==5.11.0
prettytable==3.5.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
python-cinderclient==9.1.0
python-keystoneclient==5.0.1
python-novaclient==18.2.0
python-openstackclient==6.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requestsexceptions==1.4.0
rfc3986==2.0.0
shortuuid==1.0.11
simplejson==3.18.0
six==1.16.0
sshtunnel==0.4.0
stevedore==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.11.0
